+++ PSPocs/PSPocs_Default.ps1
@@ -8,6 +8,8 @@
 $default_library = @{ 
     "name-of-library" =  @{
         "type" = "type"
+        "name" = "name"
+        "description" = "description"
         "dir" = "path-to-library"
         "library-x" = "path-to-library-x"
         "use-shared-folders" = "true"
@@ -16,5 +18,6 @@ $default_library = @{
         "papis-dir" = $(Get-ConfigProjectDir -Name 'papis')
         "local-config-file" = "%(papis-dir)s\name-of-library\name-of-library.ini"
         "cache-dir" = "%(papis-dir)s\name-of-library"
+        "url" = "url"
     }
 }